Simplify and secure
