+++ packages/gitbook-v2/src/lib/data/index.ts
@@ -5,5 +5,4 @@ export * from './errors';
 export * from './lookup';
 export * from './visitor';
 export * from './pages';
-export * from './files';
-export * from './reusable-contents';
+export * from './revisions';
