+++ packages/gitbook/src/lib/v1.ts
@@ -21,12 +21,9 @@ import {
     getEmbedByUrlInSpace,
     getLatestOpenAPISpecVersionContent,
     getPublishedContentSite,
-    getReusableContent,
     getRevision,
-    getRevisionFile,
     getRevisionPageByPath,
     getRevisionPageDocument,
-    getRevisionPages,
     getSiteRedirectBySource,
     getSpace,
     getUserById,
@@ -185,23 +182,6 @@ function getDataFetcherV1(apiTokenOverride?: string): GitBookDataFetcher {
             );
         },
 
-        getRevisionFile(params) {
-            return withAPI(() =>
-                wrapDataFetcherError(async () => {
-                    const revisionFile = await getRevisionFile(
-                        params.spaceId,
-                        params.revisionId,
-                        params.fileId
-                    );
-                    if (!revisionFile) {
-                        throw new DataFetcherError('Revision file not found', 404);
-                    }
-
-                    return revisionFile;
-                })
-            );
-        },
-
         getRevisionPageMarkdown() {
             throw new Error('Not implemented in v1');
         },
@@ -232,16 +212,6 @@ function getDataFetcherV1(apiTokenOverride?: string): GitBookDataFetcher {
             );
         },
 
-        getRevisionPages(params) {
-            return withAPI(() =>
-                wrapDataFetcherError(async () => {
-                    return getRevisionPages(params.spaceId, params.revisionId, {
-                        metadata: params.metadata,
-                    });
-                })
-            );
-        },
-
         getRevisionPageDocument(params) {
             return withAPI(() =>
                 wrapDataFetcherError(async () => {
@@ -268,24 +238,6 @@ function getDataFetcherV1(apiTokenOverride?: string): GitBookDataFetcher {
                     }
 
                     return revisionPage;
-                })
-            );
-        },
-
-        getReusableContent(params) {
-            return withAPI(() =>
-                wrapDataFetcherError(async () => {
-                    const reusableContent = await getReusableContent(
-                        params.spaceId,
-                        params.revisionId,
-                        params.reusableContentId
-                    );
-
-                    if (!reusableContent) {
-                        throw new DataFetcherError('Reusable content not found', 404);
-                    }
-
-                    return reusableContent;
                 })
             );
         },
