+++ packages/gitbook/src/routes/llms-full.ts
@@ -5,7 +5,7 @@ import { getSiteStructureSections } from '@/lib/sites';
 import { checkIsAnchor, checkIsExternalURL } from '@/lib/urls';
 import type { RevisionPageDocument, SiteSection, SiteSpace } from '@gitbook/api';
 import { type GitBookSiteContext, checkIsRootSiteContext } from '@v2/lib/context';
-import { throwIfDataError } from '@v2/lib/data';
+import { getSpaceRevision, throwIfDataError } from '@v2/lib/data';
 import assertNever from 'assert-never';
 import type { Link, Paragraph, Root } from 'mdast';
 import { fromMarkdown } from 'mdast-util-from-markdown';
@@ -101,14 +101,13 @@ async function streamMarkdownFromSiteSpaces(
         if (!siteSpaceUrl) {
             continue;
         }
-        const rootPages = await throwIfDataError(
-            dataFetcher.getRevisionPages({
-                spaceId: siteSpace.space.id,
+        const revision = await throwIfDataError(
+            getSpaceRevision(dataFetcher, {
+                space: siteSpace.space,
                 revisionId: siteSpace.space.revision,
-                metadata: false,
             })
         );
-        const pages = getIndexablePages(rootPages);
+        const pages = getIndexablePages(revision.pages);
 
         for await (const markdown of pMapIterable(
             pages,
