+++ packages/gitbook/src/routes/llms.ts
@@ -3,7 +3,7 @@ import { getIndexablePages } from '@/lib/sitemap';
 import { getSiteStructureSections } from '@/lib/sites';
 import type { SiteSection, SiteSpace } from '@gitbook/api';
 import { type GitBookSiteContext, checkIsRootSiteContext } from '@v2/lib/context';
-import { throwIfDataError } from '@v2/lib/data';
+import { getSpaceRevision, throwIfDataError } from '@v2/lib/data';
 import assertNever from 'assert-never';
 import type { ListItem, Paragraph, Root, RootContent } from 'mdast';
 import { toMarkdown } from 'mdast-util-to-markdown';
@@ -133,14 +133,13 @@ async function getNodesFromSiteSpaces(
             if (!siteSpaceUrl) {
                 return [];
             }
-            const rootPages = await throwIfDataError(
-                dataFetcher.getRevisionPages({
-                    spaceId: siteSpace.space.id,
+            const revision = await throwIfDataError(
+                getSpaceRevision(dataFetcher, {
+                    space: siteSpace.space,
                     revisionId: siteSpace.space.revision,
-                    metadata: false,
                 })
             );
-            const pages = getIndexablePages(rootPages);
+            const pages = getIndexablePages(revision.pages);
             const listChildren = await Promise.all(
                 pages.map(async ({ page }): Promise<ListItem> => {
                     const pageURL = new URL(siteSpaceUrl);
